Summarize current plant specimen records at REFLORA Virtual Herbarium

Summarize current available plant specimen records at REFLORA Virtual Herbarium hosted by the Rio de Janeiro Botanical Garden.

Usage

reflora_summary(herbarium = NULL,
                verbose = TRUE,
                save = TRUE,
                dir = "reflora_summary")

Arguments

herbarium

A vector of specific herbarium acronyms (collection code) in uppercase letters or leave it as NULL to summarize specimen records for all REFLORA-hosted herbaria.

verbose

Logical, if FALSE, a message showing steps when summarizing specimen records will not be printed in the console in full.

save

Logical, if TRUE, the search results will be saved on disk.

dir

Pathway to the computer's directory, where the table-formatted summary will be saved. The default is to create a directory named reflora_summary.

Value

A dataframe summarizing current available plant specimen records in REFLORA Virtual Herbarium.

See also

Author

Domingos Cardoso

Examples

if (FALSE) { # \dontrun{

reflora_summary(herbarium = c("ALCB", "HUEFS", "K", "RB"),
                verbose = TRUE,
                save = TRUE,
                dir = "reflora_summary")
} # }